Implement `Sanitizer.getDefaultConfiguration()`
> The sanitizer has a built-in default configuration, which is stricter than the baseline and aims to eliminate any script-injection possibility, as well as legacy or unusual constructs.

Currently we are returning the baseline `allowElements` and `allowAttributes` lists instead of the [stricter default configuration](https://wicg.github.io/sanitizer-api/#ref-for-default-configuration%E2%91%A5).

> If a Sanitizer has not received an explicit configuration, for example when being constructed without any parameters, then the default configuration value is used as the configuration object.

When the user doesn't specify an explicit config, we are currently returning a config with each configuration value set to the baseline. Now, a user who hasn't specified a explicit config will have the default configuration set for them.

https://wicg.github.io/sanitizer-api/#config